+++ R/sjPlotGLME.R
@@ -270,8 +270,7 @@ sjp.glmer <- function(fit,
                       prnt.plot = TRUE,
                       ...) {
 
-  if (stats::runif(1) < .2)
-    message("`sjp.glmer()` will become deprecated in the future. Please use `plot_model()` instead.")
+  .Deprecated("plot_model")
 
   # -------------------------------------
   # check for deprecated argument values
@@ -575,8 +574,7 @@ sjp.lmer <- function(fit,
                      prnt.plot = TRUE,
                      ...) {
 
-  if (stats::runif(1) < .2)
-    message("`sjp.lmer()` will become deprecated in the future. Please use `plot_model()` instead.")
+  .Deprecated("plot_model")
 
   # -------------------------------------
   # check for deprecated argument values