+++ R/sjPlotLinreg.R
@@ -297,8 +297,7 @@ sjp.lm <- function(fit,
                    prnt.plot = TRUE,
                    ...) {
 
-  if (stats::runif(1) < .2)
-    message("`sjp.lm()` will become deprecated in the future. Please use `plot_model()` instead.")
+  .Deprecated("plot_model")
 
   # -----------------------------------------------------------
   # remember length of predictor variables